Type: Split testing of ast and type modules.
